feat(cli): add actions query-tab-name and query-pane-name
Add the CLI actions query-tab-name and query-pane-name to enable
programmatically accessing the current active tab/pane name.

Testing: locally tested to a rather limited extent; I think the `route`
functions for the test and for the actual CLI is different in terms of
the client ID, which does happen to be a particularly weird point in
this feature, so I am uncertain as to how the unit testing should proceed.